alternative wording against confusion
when i was solving this exercise it wasn't clear to me, that `prices` has exclusively tshirts, which led me to think that `name` was not a unique field. this is why my `$match` pipeline stage looked like this:
```js
$match: {
            name: "Tshirt"
            size: process.argv[2]
        }
    };
```
which turned out being (of course) wrong. But in your wording I still think it would be possible to have more than Tshirts in the collection (e.g. Sweatshirts)
Thanks for making this great course!