Correcting Resource URL For Count Exercise (#39)
* Resource Correction

Corrected the URL in the resource section. It was previously linking to count on the "Aggregation Commands." I changed the URL to the db.collections.count() method documentation. Should help clear up confusion for newbies. PS. This is my first ever commit request. Let me know if I did anything wrong!

* Resource Correction

Same as previous correction just took off the ending part of url after the # so that it loads to top of page.